Rewrote several functions
applyBoxcarWidth: no longer uses built-in function colMeans(). Instead,
we use a simple for loop to calculate an average of the given interval.
This function appears to complete much faster.

peakr: improved readability by renaming some internal variables. Rewrote
the return call using built-ins instead of a for loop.

maxAbsFilter: this function now computes and sorts extrema on its own
instead of having an "indices" vector passed as an argument.
Additionally, this function takes a new argument "type" which allows for
different types of filtering.

includeRows: corrected an error when sorting pointList (xCol does not
necessarily have a value of 1,  and the first column of pointList stores
wavelengths by construction)

matchRowToLit: removed unused and uncessary internal variables. This
function now calls extrapolateV automatically. (This could possibly be
undesirable. However, the most common error--double assignment--should
be caught by extrapolateV.)

extrapolateV: this function no longer assumes it's dataFrame argument is
sorted; instead, it explicitly sorts dataFrame.

This commit should not be specific to "cleaner data", and should work
as-is in the master branch.